improvement(overhaul): Added tests, poetry integration, and renamed library to KinPy
The aim of the commits was to add standard library elements to the repository. Poetry is more deterministic than setuptools; therefore, more reliable. The name was changed to emphasize that the library can function for frameworks other than DLC.

fix(compute.angle): Angle was calculated incorrectly, switching to arctan2 solution

The previous method for calculating clockwise angle was wrong. The new strategy uses a more streamlined approach using arctan2, which is unit-circle-quadrant agnostic unlike regular arctan, and overcomes the limitations that would prohibit its use. Refer to angle.py for more details (will be implemented in upcoming commmit).